pyproject.toml
[tool.poetry]
name = "py-dagger"
packages = [{ include = "dagger" }]
description = "Define sophisticated data pipelines with Python and run them on different distributed systems (such as Argo Workflows)."
authors = ["larribas <lorenzo.s.arribas@gmail.com>"]
version = "0.0.0"
license = "Apache-2.0"
readme = "README.md"
repository = "https://github.com/larribas/dagger"
documentation = "https://larribas.me/dagger"
keywords = [
  "argo-workflows",
  "data-engineering",
  "data-pipelines",
  "data-science"
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Natural Language :: English",
    "Programming Language :: Python :: 3 :: Only",
    "Topic :: Software Development :: Build Tools"
]

[tool.poetry.dependencies]
python = ">=3.8,<4.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^20.8b1"
isort = "^5.7.0"
deepdiff = "^5.2.3"
flake8 = "^3.9.2"
pydocstyle = "^6.1.1"
mypy = "^0.812"
pytest-cov = "^2.12.0"
PyYAML = "^5.4.1"
mkdocs = "^1.2.2"
mkapi = "^1.0.14"
mkdocs-material = "^7.2.6"
pyspelling = "^2.7.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.pydocstyle]
inherit = false
convention = "numpy"

